+++ src/main/java/br/com/fiap/global/baze/repository/RegistroRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.fiap.global.baze.model.Registro;
 
-public interface RegistroRepository extends JpaRepository<Registro, Integer> {
+public interface RegistroRepository extends JpaRepository<Registro, Long> {
     
 }
